backend/requirements.txt
pipecat-ai[silero,deepgram,elevenlabs,anthropic]==0.0.54
twilio==9.0.4
deepgram-sdk==3.4.0
anthropic==0.39.0
elevenlabs==1.9.0
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9
supabase==2.9.1
aiofiles==24.1.0
google-api-python-client==2.140.0
google-auth-oauthlib==1.2.1
